requirements.txt
discord.py
discord.py[voice]
aiohttp
websockets
chardet